requirements.txt
flask
flask_login
Werkzeug==2.2.3
flask_mysql_connector
pymysql
mysql-connector-python
python-dotenv